Fix #68 (tmp project becomes default)
Change from `board.Save` to `pcbnew.SaveBoard` so `aSkipSettings` can be set to `True`. The code path is the same, ultimately, but with `aSkipSettings` set to `True` the bit where `SETTINGS_MANAGER` is invoked to "Set the currently loaded project path and save it" is skipped. The net effect is that file save is the same, but the user's recent projects history is not changed. So the tmp project does not appear in the "Open Recent" list in the File menu, and KiCad does not restart with the tmp project by default.